Implement inner product info filter
